backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
python-dotenv
langdetect
requests
pydantic-settings
openai
alembic
slowapi
pypdf
python-docx
python-multipart